feat: 테스트 평가 기준 컴포넌트 구현
